refactor: lift state up into TasksDashboard
Fixes #77
